Fix: non-vanilla custom blockstate registration (#5310)
* Try fixing custom blockstate registration

* another attempt; round two

* register collisions

* don't register block states twice
